Added nw condition polling & control packets

+++ cluster_topo.py
@@ -15,8 +15,8 @@ import subprocess
 import os
 import matplotlib.pyplot as plt
 from cluster_load_balancer import SimpleSwitch13
-from scapy.all import IP, IPOption, send, PcapReader, rdpcap
-
+from scapy.all import IP, TCP, send, sendp, rdpcap, Ether
+import re
 
 # Where to read/write experiment results
 RYU_LOGFILE = './ryu.log'
@@ -95,6 +95,8 @@ class ClusterTopo( Topo ):
                 break
         p = self.net.get('pgen').popen(f"sudo ifconfig pgen-eth0 mtu 65535")
         p = self.net.get('ctgen').popen(f"sudo ifconfig ctgen-eth0 mtu 65535")
+        self.poll_flag = False
+        self.cross_traffic_flag = False
         for i, host in enumerate(self.hosts):
             host.cmd(f"sudo ifconfig h{i+1}-eth0 mtu 65535")
         else:
@@ -112,6 +114,7 @@ class ClusterTopo( Topo ):
             print('\t2: Generate traffic (deprecated)')
             print('\t3: Open xterms for h1, ctgen, pgen')
             print('\t4: Run experiment')
+            print('\t5: Test')
             try:
                 choice = input()
             except EOFError:
@@ -126,6 +129,8 @@ class ClusterTopo( Topo ):
                 CLI(self.net, stdin=sys.stdin, script="./open_xterms.sh")
             elif choice == '4':
                 self.run_experiment()
+            elif choice == '5':
+                print(self.network_condition_polling())
             else:
                 print(choice, 'is not a valid option')
 
@@ -156,20 +161,17 @@ class ClusterTopo( Topo ):
                 self.stage_two_exp()
 
     def clear_experiment_logs(self):
-        """Helper function to clear all logs by executing commands on c0.
-        Preserves header line in experiment.csv file"""
-        # self.net.get('c0').cmd(f"sed -i '2,$d' ./experiment.csv", shell=True)
-        # self.net.get('c0').cmd(f"sed -i '1,$d' ./ryu.log", shell=True)
+        """Helper function to clear all logs"""
         for log in ["./ryu.log", "./experiment.csv"]:
-            # self.net.get('c0').cmd(f"truncate -s 0 {log}")
             with open(log, 'w') as f:
                 f.write('')
 
     def stage_one_exp(self):
         """Build a confusion matrix for a given PCAP validation file"""
-        results = dict()
+        results = {'cpu': dict(), 'network': dict(), 'memory': dict()}
         self.clear_experiment_logs()
-        start_time = time.time()
+        print("Running 1st stage experiment")
+        print(f"Cross traffic enabled: {self.cfg['cross_traffic']['enabled']}")
         if self.cfg["cross_traffic"]["enabled"]:
             print(f"Cross traffic enabled in mode {self.cfg['cross_traffic']['mode']}")
             p1 = multiprocessing.Process(target = self.generate_cross_traffic)
@@ -178,68 +180,39 @@ class ClusterTopo( Topo ):
                 time.sleep(1)
             elif self.cfg["cross_traffic"]["mode"] == "after":
                 p1.start(delay=1)
-        print(f"Cross traffic enabled: {self.cfg['cross_traffic']['enabled']}")
-        print(f"Experiment started at {start_time}")
-        # Value to track where to start in the CSV
-        end_of_exp = {'cpu': 2089, 'network': 4325, 'memory': 5971}
-        tmp_line_num = 1
-        p = self.net.get('pgen').cmd(
+        p = self.net.get('pgen').popen(
             ['python3', './packet_generator.py', 
              "-f", self.cfg["first_stage_experiment_files"]["cpu"],
              "-f", self.cfg["first_stage_experiment_files"]["network"],
              "-f", self.cfg["first_stage_experiment_files"]["memory"],
-             '-c', str(self.cfg["experiment_packet_count"])])
-        # for exp, fname in self.cfg["first_stage_experiment_files"].items():
-        #     # Clear the ryu log
-        #     print(f"Running {exp} experiment from {fname}")
-        #     results[exp] = dict()
-        #     if not os.path.exists(fname):
-        #         print("ERROR: Could not find file", fname)
-        #         continue
-            # end_of_exp[exp] = int(p.split('\n')[-2]) + tmp_line_num
-            # tmp_line_num += int(p.split('\n')[-2])
-            # for prediction_result in ["cpu", "network", "memory"]:
-            #     p = subprocess.Popen(
-            #         f'grep "PREDICTION: {prediction_result}" ./ryu.log | wc -l',
-            #         shell=True, stdout=subprocess.PIPE)
-            #     results[exp][prediction_result] = int(p.communicate()[0])
-        if not os.path.exists(self.cfg["experiment_results"]):
-            print("[ERROR] Could not find experiment results")
-            return
-        with open(self.cfg["experiment_results"], 'r') as f:
-            controller_results = f.readlines()
-            max_wait = 200
-            i = 0
-            while len(controller_results) < end_of_exp['memory']:
-                i += 1
-                if i == max_wait:
-                    print("Timed out")
-                    return
-                print(f"Experiment file not finished (on line {len(controller_results)}), buffering for 1s")
-                time.sleep(1)
-                controller_results.extend(f.readlines())
-            if self.cfg["cross_traffic"]["enabled"]:
-                p1.terminate()
-                self.net.get('ctgen').cmd('pkill iperf3',shell=True)
-        #     print(len(controller_results))
-        #     # print(f"Time for experiment: {time.time()-start_time}s")
-        #     # print(f"Average packet handling time: {1000*(time.time()-start_time)/len(controller_results)}ms")
-        #     print(end_of_exp)
-        #     for prediction_result in ["cpu", "network", "memory"]:
-        #         results['cpu'][prediction_result] = len([row for row in controller_results[1:end_of_exp['cpu']] if row.split(",")[0] == prediction_result])
-        #         results['network'][prediction_result] = len([row for row in controller_results[end_of_exp['cpu']:end_of_exp['network']] if row.split(",")[0] == prediction_result])
-        #         results['memory'][prediction_result] = len([row for row in controller_results[end_of_exp['network']:end_of_exp['memory']] if row.split(",")[0] == prediction_result])
-        # print(pd.DataFrame.from_dict(results).to_markdown())
-        # total_count = 0
-        # total_correct = 0
-        # for exp in ["cpu", "memory", "network"]:
-        #     for label in ["cpu", "memory", "network"]:
-        #         total_count += results[exp][label]
-        #         if exp == label:
-        #             total_correct += results[exp][label]
-        # print("Correct:", total_correct)
-        # print("Total count:", total_count)
-        # print(f"Total Accuracy: {100*total_correct/total_count}%")
+             '-c', str(self.cfg["experiment_packet_count"])], stdout=subprocess.PIPE)
+        output = p.communicate()[0].decode('utf-8').split('\n')[-2]
+        packet_list_sizes = [int(v) for v in output.split(',')]
+        # Value to track where to start in the CSV
+        end_of_exp = {
+            'cpu': packet_list_sizes[0],
+            'network': sum(packet_list_sizes[0:2]),
+            'memory': sum(packet_list_sizes)}
+        print(end_of_exp)
+        # Get the workload prediction for each packet in experiment.csv
+        prediction_results = [
+            row[0] for row in self.get_experiment_results(
+            self.cfg["experiment_results"], end_of_exp['memory'])]
+        for prediction_result in ["cpu", "network", "memory"]:
+            results['cpu'][prediction_result] = prediction_results[1:end_of_exp['cpu']].count(prediction_result)
+            results['network'][prediction_result] = prediction_results[end_of_exp['cpu']:end_of_exp['network']].count(prediction_result)
+            results['memory'][prediction_result] = prediction_results[end_of_exp['network']:end_of_exp['memory']].count(prediction_result)
+        print(pd.DataFrame.from_dict(results).to_markdown())
+        total_count = 0
+        total_correct = 0
+        for exp in ["cpu", "memory", "network"]:
+            for label in ["cpu", "memory", "network"]:
+                total_count += results[exp][label]
+                if exp == label:
+                    total_correct += results[exp][label]
+        print("Correct:", total_correct)
+        print("Total count:", total_count)
+        print(f"Total Accuracy: {100*total_correct/total_count}%")
         if self.cfg["cross_traffic"]["enabled"]:
             p1.terminate()
             self.net.get('ctgen').cmd('pkill -9 iperf3',shell=True)
@@ -248,11 +221,14 @@ class ClusterTopo( Topo ):
     def stage_two_exp(self):
         """Report average latency, bandwidth, utilization, etc.
         for each server"""
-        # Clear the ryu log
-        # with open(RYU_LOGFILE, 'w') as f:
-        #     f.write('')
-        # self.clear_experiment_logs()
-        print(f"Running 2nd stage test")
+        self.clear_experiment_logs()
+        print(f"Running 2nd stage experiment")
+        print(f"Live bandwidth polling enabled: {self.cfg['live_bandwidth_polling']['enabled']}")
+        print(f"Cross traffic enabled: {self.cfg['cross_traffic']['enabled']}")
+        if self.cfg["live_bandwidth_polling"]["enabled"]:
+            p2 = multiprocessing.Process(target=self.network_condition_polling)
+            self.poll_flag = True
+            p2.start()
         if self.cfg["cross_traffic"]["enabled"]:
             print(f"Cross traffic enabled in mode {self.cfg['cross_traffic']['mode']}")
             p1 = multiprocessing.Process(target = self.generate_cross_traffic)
@@ -261,29 +237,43 @@ class ClusterTopo( Topo ):
                 time.sleep(1)
             elif self.cfg["cross_traffic"]["mode"] == "after":
                 p1.start(delay=1)
-            p = self.net.get('pgen').popen(['python3', './packet_generator.py', "-f", self.cfg["second_stage_experiment_file"], '-c', str(self.cfg["experiment_packet_count"])])
-            p.communicate()
+        p = self.net.get('pgen').popen(
+            ['python3', './packet_generator.py', "-f",
+             self.cfg["second_stage_experiment_file"], '-c',
+             str(self.cfg["experiment_packet_count"])], stdout=subprocess.PIPE)
+        output = p.communicate()[0].decode('utf-8').split('\n')[-2]
+        print(output)
+        total_packets = sum(int(v) for v in output.split(','))
+        results = self.get_experiment_results(self.cfg["experiment_results"], total_packets)
+        if self.cfg["live_bandwidth_polling"]["enabled"]:
+            print("Terminating control packets")
+            self.poll_flag = False
+            p2.terminate()
+        if self.cfg["cross_traffic"]["enabled"]:
+            print("Terminating cross traffic")
             p1.terminate()
-            self.net.get('ctgen').cmd('pkill iperf3',shell=True)
-        else:
-            p = self.net.get('pgen').popen(['python3', './packet_generator.py', "-f", self.cfg["second_stage_experiment_file"], '-c', str(self.cfg["experiment_packet_count"])])
-            p.communicate()
+            self.net.get('ctgen').cmd('pkill -9 iperf3',shell=True)
+
+    def get_experiment_results(
+            self, file, expected_lines, max_wait_seconds=200,
+            buffer_time_seconds=1):
+        """Helper function which returns whenever the experiment.csv file
+        has all the lines needed according to the validation set sizing.
+        Returns a list of the experiment results"""
         if not os.path.exists(self.cfg["experiment_results"]):
             print("[ERROR] Could not find experiment results")
             return
-        results = csv.DictReader(open(self.cfg["experiment_results"], 'r'))
-        latencies = [int(result["latency"].replace("ms", "")) for result in results]
-        bandwidths = [int(result["bandwidth"]) for result in results]
-        print(latencies)
-        print(bandwidths)
-        if len(latencies) > 0:
-            print("Average latency:", sum(latencies)/len(latencies))
-        else:
-            print("Could not retrieve latencies")
-        if len(bandwidths) > 0:
-            print("Average bandwidth:", sum(bandwidths)/len(bandwidths))
-        else:
-            print("Could not retrieve bandwidths")
+        with open(file, 'r') as f:
+            results = f.readlines()
+            i = 0
+            while len(results) < expected_lines:
+                i += 1
+                if i == max_wait_seconds/buffer_time_seconds:
+                    print("Timed out")
+                    return
+                print(f"Experiment file not finished (on line {len(results)}), buffering for 1s")
+                time.sleep(1)
+                results.extend(f.readlines())
 
     def quit(self):
         """Necessary to let Mininet clean itself for next run"""
@@ -291,18 +281,28 @@ class ClusterTopo( Topo ):
         self.net.stop()
         exit()
 
-
     def generate_cross_traffic(self, delay=0):
         print("Setting up iperf servers on machines")
         for i, host in enumerate(self.hosts):
             p = host.popen(f"iperf3 -s -D")
             print(p.communicate())
         ip_list = [f'10.0.0.{i}' for i in range(1, self.k + 1)]
-        time.sleep(delay)
+        # Clear out any logs
         for ip in ip_list:
-            # os.remove(f'./cross_traffic_log_{ip}.log')
             with open(f'./cross_traffic_log_{ip}.log', 'w') as f:
                 f.write('')
+        time.sleep(delay)
+        if self.cfg['cross_traffic']['parallel_machines'] < self.k + 1:
+            while self.cross_traffic_flag:
+                for ip in random.sample(ip_list, self.cfg['cross_traffic']['parallel_machines']):
+                    iperf_cmd = f'''\
+iperf3 -i {self.cfg["cross_traffic"]["logging_interval"]} \
+-c {ip} -u -t {self.cfg["cross_traffic"]["parallel_switching_interval"]} \
+--logfile ./cross_traffic_log_{ip}.log \
+--zerocopy -l {self.cfg["cross_traffic"]["buffer_length"]} -w 2048'''
+                    print(iperf_cmd)
+                    p = self.net.get('ctgen').popen(iperf_cmd, shell=True)
+                p.communicate()
         for ip in ip_list:
             iperf_cmd = f'''\
 iperf3 -i {self.cfg["cross_traffic"]["logging_interval"]} \
@@ -313,6 +313,62 @@ iperf3 -i {self.cfg["cross_traffic"]["logging_interval"]} \
             p = self.net.get('ctgen').popen(iperf_cmd, shell=True)
         return
 
+    def network_condition_polling(self):
+        """Process to manage all polls, including wait times, and sending the
+        control packets to the controller"""
+        interval = self.cfg["live_bandwidth_polling"]["interval"]
+        while self.poll_flag:
+            bandwidths = self.check_bandwidths()
+            round_trip_times, packetlosses = self.check_rtt_packetloss()
+            if len(bandwidths) != 48:
+                print(f"ERROR: {bandwidths}")
+            if len(round_trip_times) != 48:
+                print(f"ERROR: {round_trip_times}")
+            if len(packetlosses) != 48:
+                print(f"ERROR: {packetlosses}")
+            all_data = b''.join([bandwidths, round_trip_times, packetlosses])
+            print(all_data)
+            cmd = f"scapy -H << here\nsend(IP(dst='10.0.0.111')/ICMP()/{str(all_data).replace('`', '')})\nhere"
+            print(cmd)
+            p = self.net.get('pgen').popen(cmd, shell=True)
+            print(p.communicate())
+            time.sleep(interval)
+
+    def check_bandwidths(self):
+        """Returns a list of the bandwidths of each interface. Returns a string
+        of bytes to be bundled into a control packet"""
+        print("Checking Bandwidths")
+        commands = []
+        for i, host in enumerate(self.hosts):
+            commands.append(host.popen(f"./bwmonitor h{i+1}-eth0 0.1 0.1", shell=True, stdout=subprocess.PIPE))
+        bandwidths = []
+        for command in commands:
+            tmp = (command.communicate()[0].decode('utf-8').strip())
+            tmp = int(tmp) if tmp != '' else 0
+            bandwidths.append(tmp)
+        bandwidths = b''.join([int.to_bytes(bw, 4, 'big') for bw in bandwidths])
+        return bandwidths
+
+    def check_rtt_packetloss(self):
+        """Uses the ping command to calculate rtt and packetloss for each host
+        in the topology. Returns a string of bytes to be bundled into a control
+        packet."""
+        print("Checking rtt and packetloss")
+        commands = []
+        for i in range(1, len(self.hosts) + 1):
+            ping_cmd = f'ping 10.0.0.{i} -c {self.cfg["live_bandwidth_polling"]["ping_counts"]}'.split(' ')
+            commands.append(self.net.get('pgen').popen(ping_cmd, stdout=subprocess.PIPE))
+        round_trip_times = []
+        packetlosses = []
+        for command in commands:
+            output = command.communicate()[0].decode('utf-8').split('\n')
+            # Convert round_trip_times to microseconds, turn into integer for transmit
+            round_trip_times.append(int(float(output[-2].split(' ')[3].split('/')[0])*1000))
+            # Convert packetlosses to packets per million for transmit
+            packetlosses.append(int(float(output[-3].split(' ')[5].replace('%', ''))*1000000))
+        round_trip_times = b''.join([int.to_bytes(lt, 4, 'big') for lt in round_trip_times])
+        packetlosses = b''.join([int.to_bytes(pl, 4, 'big') for pl in packetlosses])
+        return round_trip_times, packetlosses
 
 if __name__ == "__main__":
     topo = ClusterTopo()
